refactor: Grouped related files (create-signatures)
