+++ src/steps/create-signatures/create-signature/index.ts
@@ -1,0 +1,3 @@
+export * from './pass-signature-to-base-component.js';
+export * from './update-constructor.js';
+export * from './update-reference.js';
