+++ src/steps/create-signatures/create-signature/pass-signature-to-base-component.ts
@@ -1,6 +1,6 @@
 import { AST } from '@codemod-utils/ast-javascript';
 
-import type { TransformedEntityName } from '../../utils/components.js';
+import type { TransformedEntityName } from '../../../utils/components.js';
 import {
   builderConvertArgsToSignature,
   builderCreateSignature,
